[Python] Move BatchElementExceptions into errors
Move BatchElementErrors into the errors namespace.
Split BatchElementError from BatchElementException, and make the binding
cpp structures mirror eachother.
