+++ src/openassetio-python/cmodule/src/_openassetio.hpp
@@ -113,10 +113,10 @@ void registerManagerStateBase(const py::module& mod);
 void registerEntityReference(const py::module& mod);
 
 /// Register the BatchElementError type with Python.
-void registerBatchElementError(py::module& mod);
+void registerBatchElementError(const py::module& mod);
 
-// Register errors, linking exceptions to/from Python.
-void registerErrors(const py::module& mod);
+// Register exceptions, including BatchElementExceptions.
+void registerExceptions(const py::module& mod);
 
 /// Register the EntityReferencePager class with Python.
 void registerEntityReferencePager(const py::module& mod);
