+++ src/openassetio-python/tests/package/test_batchelementerror.py
@@ -23,7 +23,7 @@ Tests for the BatchElementError type.
 
 import pytest
 
-from openassetio import (
+from openassetio.errors import (
     BatchElementError,
     BatchElementException,
     UnknownBatchElementException,
